added some packages to organize the classes and updated search function
