+++ App/app/src/main/java/com/example/barterbooksapp/CategoriesActivity.java
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import android.os.Bundle;
 
+import com.example.barterbooksapp.recyclerViewAdapters.LocCatListRecyclerView;
 import com.google.firebase.auth.FirebaseAuth;
 
 import java.util.ArrayList;
