+++ App/app/src/main/java/com/example/barterbooksapp/LocationActivity.java
@@ -7,6 +7,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import android.content.Intent;
 import android.os.Bundle;
 
+import com.example.barterbooksapp.recyclerViewAdapters.LocCatListRecyclerView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
