+++ App/app/src/main/java/com/example/barterbooksapp/recyclerViewAdapters/LocCatListRecyclerView.java
@@ -1,4 +1,4 @@
-package com.example.barterbooksapp;
+package com.example.barterbooksapp.recyclerViewAdapters;
 
 import android.content.Context;
 import android.content.Intent;
@@ -7,10 +7,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.barterbooksapp.MainActivity;
+import com.example.barterbooksapp.R;
 
 import java.util.List;
