+++ App/app/src/main/java/com/example/barterbooksapp/recyclerViewAdapters/MyRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.barterbooksapp;
+package com.example.barterbooksapp.recyclerViewAdapters;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -12,6 +12,10 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.barterbooksapp.DetailsActivity;
+import com.example.barterbooksapp.R;
+import com.example.barterbooksapp.utlity.BookPostDataModel;
 
 import java.util.List;
