+++ App/app/src/main/java/com/example/barterbooksapp/utlity/DBUtilities.java
@@ -1,4 +1,4 @@
-package com.example.barterbooksapp;
+package com.example.barterbooksapp.utlity;
 
 import android.util.Log;
